deleted unused ui parameters
mvoed the obstacle course demo to the UI
