--- src/main/java/us/ihmc/valkyrie/parameters/ValkyrieUIFootstepGeneratorParameters.java
@@ -1,11 +1,0 @@
-package us.ihmc.valkyrie.parameters;
-
-import us.ihmc.humanoidRobotics.footstep.footstepGenerator.UIFootstepGeneratorParameters;
-
-public class ValkyrieUIFootstepGeneratorParameters extends UIFootstepGeneratorParameters
-{
-   public ValkyrieUIFootstepGeneratorParameters()
-   {
-      turningStepWidth = 0.30;
-   }
-}
