--- src/test/java/us/ihmc/valkyrie/obstacleCourse/ValkyrieObstacleCourseRampFootstepSnapperTest.java
@@ -1,33 +1,0 @@
-package us.ihmc.valkyrie.obstacleCourse;
-
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-
-import us.ihmc.avatar.drcRobot.DRCRobotModel;
-import us.ihmc.avatar.drcRobot.RobotTarget;
-import us.ihmc.avatar.obstacleCourseTests.DRCObstacleCourseRampFootstepSnapperTest;
-import us.ihmc.simulationConstructionSetTools.tools.CITools;
-import us.ihmc.valkyrie.ValkyrieRobotModel;
-
-@Tag("humanoid-obstacle-slow")
-public class ValkyrieObstacleCourseRampFootstepSnapperTest extends DRCObstacleCourseRampFootstepSnapperTest
-{
-   @Override
-   public DRCRobotModel getRobotModel()
-   {
-      return new ValkyrieRobotModel(RobotTarget.SCS);
-   }
-
-   @Override
-   public String getSimpleRobotName()
-   {
-      return CITools.getSimpleRobotNameFor(CITools.SimpleRobotNameKeys.VALKYRIE);
-   }
-
-   @Override
-   @Test
-   public void testWalkingUpRampUsingSnapFootsteps()
-   {
-      super.testWalkingUpRampUsingSnapFootsteps();
-   }
-}
